debt: playground: extract time-format, watchdog-ui, relay-board from main.js
Three behavior-preserving moves out of the 3556-line monolith.
Each new module takes its own state with it, exposing only the
functions main.js still needs to call.

- main/time-format.js: Helsinki-TZ formatters + cause/reason labels
  + escapeHtml (consumed by log rendering, which stays in main.js
  for now).
- main/watchdog-ui.js: banner, cool-off indicator, mode-enablement
  card, watchdogs card, history. Takes a getLiveSource callback
  so the reattach-on-liveSource-change dance works from outside.
  Drops the window.__attachWatchdogWS global.
- main/relay-board.js: manual override + relay toggles + forced-
  mode buttons. Reads watchdog snapshot via getWatchdogSnapshot().

main.js: 3468 → 2716 lines. npm test:unit 782/782 pass. knip clean.
