+++ playground/js/main.js
@@ -20,8 +20,15 @@ import { initAuth } from './auth.js';
 import { captureInstallPrompt, triggerInstall, wireInstallModal, initNotifications, subscribePush, updateCategories, unsubscribePush, isSubscribed, getSelectedCategories, sendTest } from './notifications.js';
 import { buildSchematic as buildSchematicFromSvg } from './schematic.js';
 import { SIM_START_HOUR, getDayNightEnv } from './sim-bootstrap.js';
-// connection.js actions will be used in later phases — import deferred to avoid name collisions
-// import { switchToLive, switchToSimulation, ... } from './actions/connection.js';
+import {
+  formatClockTime, formatCauseLabel, formatReasonLabel,
+  formatSensorsLine, formatFullTimeHelsinki, escapeHtml,
+} from './main/time-format.js';
+import {
+  initWatchdogUI, attachWatchdogWebSocket,
+  renderModeEnablement, getWatchdogSnapshot,
+} from './main/watchdog-ui.js';
+import { initRelayBoard, updateRelayBoard } from './main/relay-board.js';
 // Expose for e2e testing
 window.__triggerVersionCheck = triggerVersionCheck;
 
@@ -269,9 +276,7 @@ function ensureLiveSource() {
       }
     });
     // Re-attach the watchdog broadcast listener now that liveSource exists.
-    if (typeof window.__attachWatchdogWS === 'function') {
-      window.__attachWatchdogWS();
-    }
+    attachWatchdogWebSocket();
     attachScriptStatusWebSocket(liveSource);
   }
   liveSource.start();
@@ -814,8 +819,8 @@ async function init() {
   // Initialize live/simulation mode toggle
   initModeToggle();
   initDeviceConfig();
-  initRelayBoard();
-  initWatchdogUI();
+  initRelayBoard({ getLiveSource: () => liveSource });
+  initWatchdogUI({ getLiveSource: () => liveSource });
 
   // On deploys without live mode (e.g. GitHub Pages) load the
   // pre-baked simulation snapshot so the dashboard is populated on
@@ -1369,98 +1374,7 @@ function renderLogsList() {
   container.innerHTML = html;
 }
 
-// Finnish timezone everywhere in the UI. Using Intl rather than
-// getHours() which would follow the browser's locale (typically correct
-// for local operators but wrong when accessed from abroad).
-const HELSINKI_TZ = 'Europe/Helsinki';
-const fmtClockHelsinki = new Intl.DateTimeFormat('fi-FI', {
-  hour: '2-digit', minute: '2-digit', hour12: false, timeZone: HELSINKI_TZ,
-});
-const fmtFullHelsinki = new Intl.DateTimeFormat('fi-FI', {
-  year: 'numeric', month: '2-digit', day: '2-digit',
-  hour: '2-digit', minute: '2-digit', second: '2-digit',
-  hour12: false, timeZone: HELSINKI_TZ,
-});
-
-function formatClockTime(unixMs) {
-  return fmtClockHelsinki.format(new Date(unixMs));
-}
-
-// Short human-readable label for each cause tag. Keep in sync with
-// state.lastTransitionCause values set in shelly/control.js.
-const CAUSE_LABELS = {
-  boot: 'Boot',
-  automation: 'Automation',
-  forced: 'Forced mode',
-  safety_override: 'Safety override',
-  watchdog_auto: 'Watchdog (auto)',
-  user_shutdown: 'User shutdown',
-  drain_complete: 'Drain complete',
-  failed: 'Transition failed',
-};
-function formatCauseLabel(c) {
-  return CAUSE_LABELS[c] || c;
-}
-
-// Finer-grained decision codes from the evaluator (shelly/control-logic.js).
-// Paired with CAUSE_LABELS in the System Logs UI to turn a bare
-// "automation" tag into "Automation — solar stall". Keep the keys in
-// sync with the `reason` string set at each return path inside
-// evaluate().
-const REASON_LABELS = {
-  solar_enter: 'collector hot enough to charge',
-  solar_refill: 'refilling drained collectors',
-  solar_active: 'tank still gaining heat',
-  solar_stall: 'tank stopped gaining heat',
-  solar_drop_from_peak: 'tank cooling below peak',
-  overheat_circulate: 'collector overheat — circulating to cool',
-  overheat_drain: 'collector overheat — draining',
-  freeze_drain: 'freeze risk — draining',
-  drain_running: 'drain in progress',
-  drain_timeout: 'drain timeout fallback',
-  greenhouse_enter: 'greenhouse cold — heating',
-  greenhouse_active: 'greenhouse still cold',
-  greenhouse_warm: 'greenhouse warm enough',
-  greenhouse_tank_depleted: 'tank too cool to heat greenhouse',
-  emergency_enter: 'greenhouse critical — emergency heat',
-  sensor_stale: 'sensor reading stale',
-  watchdog_ban: 'mode blocked by watchdog',
-  min_duration: 'holding minimum run time',
-  idle: 'no trigger active',
-};
-function formatReasonLabel(r) {
-  return REASON_LABELS[r] || r;
-}
-
-// Render the temp snapshot as "coll 62.3° · tank 41/29° · gh 12° · out 8°"
-function formatSensorsLine(sensors) {
-  if (!sensors) return '';
-  const f = (v) => (typeof v === 'number' ? v.toFixed(1) + '°' : '—');
-  const parts = [];
-  if ('collector' in sensors) parts.push('coll ' + f(sensors.collector));
-  if ('tank_top' in sensors || 'tank_bottom' in sensors) {
-    parts.push('tank ' + f(sensors.tank_top) + '/' + f(sensors.tank_bottom));
-  }
-  if ('greenhouse' in sensors) parts.push('gh ' + f(sensors.greenhouse));
-  if ('outdoor' in sensors) parts.push('out ' + f(sensors.outdoor));
-  return escapeHtml(parts.join(' · '));
-}
-
-// YYYY-MM-DD HH:MM:SS in Europe/Helsinki — used by clipboard export.
-function formatFullTimeHelsinki(unixMs) {
-  const parts = fmtFullHelsinki.formatToParts(new Date(unixMs));
-  const get = (type) => { const p = parts.find(x => x.type === type); return p ? p.value : ''; };
-  return get('year') + '-' + get('month') + '-' + get('day') + ' ' +
-         get('hour') + ':' + get('minute') + ':' + get('second');
-}
-
-function escapeHtml(s) {
-  return String(s)
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;');
-}
+// Time formatting + cause/reason labels moved to ./main/time-format.js
 
 // Attach a one-time scroll listener that lazy-loads older events when the
 // user reaches the bottom of the System Logs card.
@@ -2741,7 +2655,7 @@ function populateDeviceForm(cfg) {
   setToggle('dc-ea-ih', !!(ea & 16));
 
   // Mode enablement card (replaces the old allowed-modes checkboxes)
-  renderModeEnablement(cfg.wb || {}, _watchdogCurrentUserRole());
+  renderModeEnablement(cfg.wb || {}, store.get('userRole') || 'admin');
 
   // Version & size
   document.getElementById('dc-version').textContent = cfg.v || '-';
@@ -2796,761 +2710,6 @@ function saveDeviceConfig() {
     });
 }
 
-// ── Relay Toggle Board ──
-let overrideActive = false;
-let overrideExpiresAt = 0;
-let overrideCountdownTimer = null;
-let relayPendingState = {}; // relay → expected state (for reconciliation)
-let relayPendingTimers = {}; // relay → timeout ID
-// Last known controls_enabled value from any state message — used for
-// optimistic re-enable of the Enter button after Exit Override, so the
-// user doesn't wait up to 30s for the next Shelly state broadcast.
-let lastControlsEnabled = false;
-// Currently active forced-mode short code (null = Automatic)
-var currentForcedMode = null;
 
-// Short code → full MODE_VALVES/MODE_ACTUATORS key
-var MODE_CODE_MAP = {
-  I: 'IDLE', SC: 'SOLAR_CHARGING', GH: 'GREENHOUSE_HEATING',
-  AD: 'ACTIVE_DRAIN', EH: 'EMERGENCY_HEATING'
-};
-
-// Original button labels keyed by data-mode value (for restoring after ban suffix)
-var FM_BTN_LABELS = {
-  I: 'Idle', SC: 'Solar charging',
-  GH: 'Greenhouse heating', AD: 'Active drain', EH: 'Emergency heating'
-};
-
-// Apply forced-mode relay preview optimistically (no server round-trip)
-function applyForcedModePreview(modeCode) {
-  if (!modeCode) return; // Automatic — let real state reconcile
-  var fullName = MODE_CODE_MAP[modeCode];
-  if (!fullName) return;
-  // loadControlLogic() is already cached after init()
-  loadControlLogic().then(function (cl) {
-    var valves = cl.MODE_VALVES[fullName] || {};
-    var actuators = cl.MODE_ACTUATORS[fullName] || {};
-    document.querySelectorAll('.relay-btn').forEach(function (btn) {
-      var relay = btn.dataset.relay;
-      var on = (relay === 'pump' || relay === 'fan')
-        ? !!actuators[relay]
-        : !!valves[relay];
-      btn.classList.toggle('on', on);
-    });
-  });
-}
-
-function initRelayBoard() {
-  // Enter override
-  document.getElementById('override-enter-btn').addEventListener('click', enterOverride);
-
-  // Exit override
-  document.getElementById('override-exit-btn').addEventListener('click', exitOverride);
-
-  // TTL buttons
-  document.querySelectorAll('.ttl-btn').forEach(btn => {
-    btn.addEventListener('click', function () {
-      document.querySelectorAll('.ttl-btn').forEach(b => b.classList.remove('active'));
-      this.classList.add('active');
-      if (overrideActive) updateOverrideTtl(parseInt(this.dataset.ttl, 10));
-    });
-  });
-
-  // Relay buttons
-  document.querySelectorAll('.relay-btn').forEach(btn => {
-    btn.addEventListener('click', function () {
-      if (this.disabled || !overrideActive) return;
-      toggleRelay(this);
-    });
-  });
-
-  // Forced-mode buttons
-  var forcedModeSendTimer = null;
-  document.querySelectorAll('#forced-mode-btns .fm-btn').forEach(btn => {
-    btn.addEventListener('click', function () {
-      if (this.disabled || !overrideActive) return;
-      var mode = this.dataset.mode;
-      if (!mode || mode === currentForcedMode) return;
-
-      document.querySelectorAll('#forced-mode-btns .fm-btn').forEach(b => b.classList.remove('active'));
-      this.classList.add('active');
-      try { if (navigator.vibrate) navigator.vibrate(50); } catch (e) {}
-      applyForcedModePreview(mode);
-
-      if (forcedModeSendTimer) clearTimeout(forcedModeSendTimer);
-      forcedModeSendTimer = setTimeout(function () {
-        if (liveSource) liveSource.sendCommand({ type: 'override-set-mode', mode: mode });
-      }, 300);
-      currentForcedMode = mode;
-    });
-  });
-
-  // Status-view "Exit override" link
-  var exitLink = document.getElementById('mode-card-exit-link');
-  if (exitLink) {
-    exitLink.addEventListener('click', function (ev) {
-      ev.preventDefault();
-      if (liveSource) liveSource.sendCommand({ type: 'override-exit' });
-    });
-  }
-
-  // Command response handler
-  if (liveSource) {
-    liveSource.onCommandResponse(handleOverrideResponse);
-  }
-}
-
-var overrideAckTimer = null;
-
-function enterOverride() {
-  if (!liveSource) return;
-  // Hard-override confirmation. Automation including freeze-drain is
-  // off while in override; the user needs to acknowledge that each
-  // time so it's not triggered by a stray click. No confirmation on
-  // subsequent fm changes within the same override session — we
-  // trust the first "yes, I meant it" for the whole TTL window.
-  var ok = window.confirm(
-    'Manual override disables ALL automation until you exit (or the TTL expires) — ' +
-    'including freeze-drain safety. On a cold night an active override can let the ' +
-    'collectors freeze.\n\nContinue?'
-  );
-  if (!ok) return;
-
-  var fmSelect = document.getElementById('override-entry-fm');
-  var fm = fmSelect ? fmSelect.value : 'I';
-
-  var btn = document.getElementById('override-enter-btn');
-  btn.disabled = true;
-  btn.textContent = 'Connecting...';
-
-  var activeTtlBtn = document.querySelector('.ttl-btn.active');
-  var ttl = activeTtlBtn ? parseInt(activeTtlBtn.dataset.ttl, 10) : 300;
-  var sent = liveSource.sendCommand({ type: 'override-enter', ttl: ttl, forcedMode: fm });
-
-  if (!sent) {
-    btn.disabled = false;
-    btn.textContent = 'Enter Manual Override';
-    showOverrideMsg('WebSocket not connected.', 'var(--error)');
-    return;
-  }
-
-  // Timeout if no ack received
-  clearTimeout(overrideAckTimer);
-  overrideAckTimer = setTimeout(function () {
-    if (!overrideActive) {
-      btn.disabled = false;
-      btn.textContent = 'Enter Manual Override';
-      showOverrideMsg('No response from server. Is the controller reachable?', 'var(--error)');
-    }
-  }, 5000);
-}
-
-function showOverrideMsg(text, color) {
-  var el = document.getElementById('override-expired-msg');
-  el.textContent = text;
-  el.style.color = color || 'var(--on-surface-variant)';
-  el.style.display = '';
-  setTimeout(function () { el.style.display = 'none'; }, 6000);
-}
-
-function exitOverride() {
-  if (!liveSource) return;
-  liveSource.sendCommand({ type: 'override-exit' });
-}
-
-function updateOverrideTtl(ttl) {
-  if (!liveSource || !overrideActive) return;
-  liveSource.sendCommand({ type: 'override-update', ttl: ttl });
-}
-
-function handleOverrideResponse(msg) {
-  clearTimeout(overrideAckTimer);
-  if (msg.type === 'override-ack') {
-    if (msg.active) {
-      activateOverrideUI(msg.expiresAt, msg.forcedMode);
-    } else {
-      deactivateOverrideUI();
-    }
-  } else if (msg.type === 'override-error') {
-    // Restore enter button
-    var btn = document.getElementById('override-enter-btn');
-    btn.disabled = false;
-    btn.textContent = 'Enter Manual Override';
-    showOverrideMsg(msg.message, 'var(--error)');
-  }
-}
-
-function activateOverrideUI(expiresAt, forcedMode) {
-  overrideActive = true;
-  overrideExpiresAt = expiresAt;
-  currentForcedMode = forcedMode || null;
-  document.getElementById('override-entry').style.display = 'none';
-  document.getElementById('override-active-header').style.display = '';
-  document.getElementById('relay-board').style.display = '';
-  document.getElementById('override-expired-msg').style.display = 'none';
-  document.querySelectorAll('.relay-btn').forEach(btn => { btn.disabled = false; });
-
-  // Show forced-mode group; gate buttons for readonly users. Reflect
-  // the active fm from the server so the button highlight matches the
-  // mode we actually entered with.
-  var fmGroup = document.getElementById('forced-mode-group');
-  if (fmGroup) fmGroup.style.display = '';
-  var userRole = store.get('userRole') || 'admin';
-  var isAdmin = userRole === 'admin';
-  document.querySelectorAll('#forced-mode-btns .fm-btn').forEach(function (b) {
-    b.disabled = !isAdmin;
-    b.classList.toggle('active', b.dataset.mode === forcedMode);
-  });
-
-  startCountdown();
-}
-
-function deactivateOverrideUI(msg) {
-  overrideActive = false;
-  overrideExpiresAt = 0;
-  clearCountdown();
-  document.getElementById('override-entry').style.display = '';
-  document.getElementById('override-active-header').style.display = 'none';
-  document.getElementById('relay-board').style.display = 'none';
-  document.querySelectorAll('.relay-btn').forEach(btn => {
-    btn.disabled = true;
-    btn.classList.remove('on', 'relay-btn--pending', 'relay-btn--error');
-  });
-  relayPendingState = {};
-  for (var k in relayPendingTimers) clearTimeout(relayPendingTimers[k]);
-  relayPendingTimers = {};
-
-  // Hide forced-mode group and reset its state
-  currentForcedMode = null;
-  var fmGroup = document.getElementById('forced-mode-group');
-  if (fmGroup) fmGroup.style.display = 'none';
-  document.querySelectorAll('#forced-mode-btns .fm-btn').forEach(function (b) {
-    b.classList.remove('active');
-    // Restore original button text (strip any " · banned" suffix)
-    var orig = FM_BTN_LABELS[b.dataset.mode !== undefined ? b.dataset.mode : ''];
-    if (orig) b.textContent = orig;
-    b.disabled = false;
-  });
-  // No Automatic tile anymore — with hard override, fm is required
-  // whenever override is active. On deactivation the whole group hides.
-  // Reset the Enter button so the user doesn't see a stale "Connecting..."
-  // and doesn't have to wait ~30s for the next state broadcast to recover.
-  var enterBtn = document.getElementById('override-enter-btn');
-  enterBtn.textContent = 'Enter Manual Override';
-  enterBtn.disabled = !lastControlsEnabled;
-  var gateMsg = document.getElementById('override-gate-msg');
-  if (gateMsg) gateMsg.style.display = lastControlsEnabled ? 'none' : 'block';
-  if (msg) {
-    var expEl = document.getElementById('override-expired-msg');
-    expEl.textContent = msg;
-    expEl.style.display = '';
-    expEl.style.color = 'var(--on-surface-variant)';
-    setTimeout(() => { expEl.style.display = 'none'; }, 5000);
-  }
-}
-
-function startCountdown() {
-  clearCountdown();
-  updateCountdownDisplay();
-  overrideCountdownTimer = setInterval(updateCountdownDisplay, 1000);
-}
-
-function clearCountdown() {
-  if (overrideCountdownTimer) { clearInterval(overrideCountdownTimer); overrideCountdownTimer = null; }
-}
-
-function updateCountdownDisplay() {
-  var remaining = Math.max(0, overrideExpiresAt - Math.floor(Date.now() / 1000));
-  var min = Math.floor(remaining / 60);
-  var sec = remaining % 60;
-  document.getElementById('override-countdown').textContent = min + ':' + (sec < 10 ? '0' : '') + sec;
-  if (remaining <= 0 && overrideActive) {
-    deactivateOverrideUI('Override expired — automation resumed.');
-  }
-}
-
-function toggleRelay(btn) {
-  var relay = btn.dataset.relay;
-  var currentlyOn = btn.classList.contains('on');
-  var newState = !currentlyOn;
-
-  // Optimistic UI + haptic feedback
-  btn.classList.toggle('on', newState);
-  btn.classList.add('relay-btn--pending');
-  try { if (navigator.vibrate) navigator.vibrate(50); } catch (e) {}
-
-  // Send command
-  if (liveSource) liveSource.sendCommand({ type: 'relay-command', relay: relay, on: newState });
-
-  // Track pending state for reconciliation
-  relayPendingState[relay] = newState;
-  if (relayPendingTimers[relay]) clearTimeout(relayPendingTimers[relay]);
-  relayPendingTimers[relay] = setTimeout(function () {
-    // Reconciliation timeout — if state hasn't been confirmed, revert
-    delete relayPendingState[relay];
-    delete relayPendingTimers[relay];
-    btn.classList.remove('relay-btn--pending');
-    // Don't revert — next state broadcast will reconcile
-  }, 2000);
-}
-
-function updateRelayBoard(result) {
-  if (!result) return;
-  var mo = result.manual_override;
-  // Update last-known controls_enabled FIRST so deactivateOverrideUI()
-  // can use the fresh value for its optimistic re-enable.
-  lastControlsEnabled = !!result.controls_enabled;
-
-  // Handle override state from server
-  if (mo && mo.active && !overrideActive) {
-    // Override started externally or on reconnect
-    activateOverrideUI(mo.expiresAt, mo.forcedMode);
-  } else if ((!mo || !mo.active) && overrideActive) {
-    // Override ended externally
-    deactivateOverrideUI('Override ended — automation resumed.');
-    return;
-  } else if (mo && mo.active && overrideActive) {
-    // Update expiry (may have been adjusted)
-    overrideExpiresAt = mo.expiresAt;
-  }
-
-  // Update controls-enabled gate
-  var ceEnabled = lastControlsEnabled;
-  var enterBtn = document.getElementById('override-enter-btn');
-  var gateMsg = document.getElementById('override-gate-msg');
-  if (!overrideActive) {
-    enterBtn.disabled = !ceEnabled;
-    gateMsg.style.display = ceEnabled ? 'none' : 'block';
-  }
-
-  // ce=false during active override → force deactivate
-  if (overrideActive && !ceEnabled) {
-    deactivateOverrideUI('Controls disabled — override ended.');
-    return;
-  }
-
-  if (!overrideActive) return;
-
-  // Update relay button states from actual hardware state
-  var valves = result.valves || {};
-  var actuators = result.actuators || {};
-  document.querySelectorAll('.relay-btn').forEach(btn => {
-    var relay = btn.dataset.relay;
-    var actual = (relay === 'pump' || relay === 'fan')
-      ? !!actuators[relay]
-      : !!valves[relay];
-
-    // Reconcile with pending state
-    if (relay in relayPendingState) {
-      if (relayPendingState[relay] === actual) {
-        // Confirmed — clear pending
-        delete relayPendingState[relay];
-        if (relayPendingTimers[relay]) { clearTimeout(relayPendingTimers[relay]); delete relayPendingTimers[relay]; }
-        btn.classList.remove('relay-btn--pending');
-      } else {
-        // State doesn't match — command may have failed
-        // Revert optimistic update
-        delete relayPendingState[relay];
-        if (relayPendingTimers[relay]) { clearTimeout(relayPendingTimers[relay]); delete relayPendingTimers[relay]; }
-        btn.classList.remove('relay-btn--pending');
-        btn.classList.add('relay-btn--error');
-        setTimeout(() => btn.classList.remove('relay-btn--error'), 400);
-      }
-    }
-
-    btn.classList.toggle('on', actual);
-  });
-
-  // Sync forced-mode button active state from server
-  var fm = (mo && mo.forcedMode) || null;
-  currentForcedMode = fm;
-  var fmCode = fm || '';
-  document.querySelectorAll('#forced-mode-btns .fm-btn').forEach(function (b) {
-    b.classList.toggle('active', b.dataset.mode === fmCode);
-  });
-
-  // Apply wb bans to forced-mode buttons
-  var wb = _watchdogSnapshot.wb || {};
-  var now = Math.floor(Date.now() / 1000);
-  document.querySelectorAll('#forced-mode-btns .fm-btn').forEach(function (b) {
-    var code = b.dataset.mode;
-    if (!code) return; // Automatic button is never banned
-    var banUntil = wb[code];
-    var isBanned = banUntil && banUntil > now;
-    var orig = FM_BTN_LABELS[code] || code;
-    if (isBanned) {
-      b.disabled = true;
-      b.textContent = orig + ' · banned';
-    } else {
-      var userRole = store.get('userRole') || 'admin';
-      b.disabled = userRole !== 'admin';
-      b.textContent = orig;
-    }
-  });
-}
-
-// ── Watchdog UI ──
-//
-// Renders three distinct pieces:
-//   1. Pending banner on #status (with live countdown + ack/shutdown)
-//   2. Cool-off inline indicator on #status
-//   3. Mode enablement card in #device (replaces old allowed-modes)
-//   4. Anomaly watchdogs card in #settings (enable/disable + snooze)
-// Live state arrives via WebSocket 'watchdog-state' messages; initial
-// state is fetched from GET /api/watchdog/state on load.
-
-const WATCHDOG_PERMANENT_SENTINEL = 9999999999;
-const WATCHDOG_ALL_MODES = [
-  { code: 'I',  label: 'IDLE' },
-  { code: 'SC', label: 'SOLAR_CHARGING' },
-  { code: 'GH', label: 'GREENHOUSE_HEATING' },
-  { code: 'AD', label: 'ACTIVE_DRAIN' },
-  { code: 'EH', label: 'EMERGENCY_HEATING' },
-];
-
-let _watchdogCountdownTimer = null;
-let _watchdogPending = null;
-let _watchdogMeta = [];
-let _watchdogSnapshot = { we: {}, wz: {}, wb: {} };
-// Tracks whether the watchdog state has been seeded by either the
-// initial GET or a WebSocket broadcast. The WS handler always wins —
-// it represents the server's most recent push — so once any state has
-// been applied, the (possibly in-flight) GET response is discarded
-// rather than allowed to clobber a fresher WS update. Without this,
-// the GET → handler order is racy: a WS broadcast that arrives
-// between the GET being issued and it resolving would be silently
-// overwritten by the older GET response.
-let _watchdogStateSeeded = false;
-
-function _watchdogCurrentUserRole() {
-  return store.get('userRole') || 'admin';
-}
-
-function initWatchdogUI() {
-  // Wire banner buttons
-  const snoozeBtn = document.getElementById('watchdog-banner-snooze');
-  const shutdownBtn = document.getElementById('watchdog-banner-shutdown');
-  const replyInput = document.getElementById('watchdog-banner-reply');
-
-  if (snoozeBtn) {
-    snoozeBtn.addEventListener('click', () => {
-      if (!_watchdogPending) return;
-      const reason = (replyInput && replyInput.value.trim()) || '(no reason provided)';
-      fetch('/api/watchdog/ack', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        credentials: 'include',
-        body: JSON.stringify({
-          id: _watchdogPending.id,
-          eventId: _watchdogPending.dbEventId,
-          reason: reason
-        })
-      }).catch(err => console.error('watchdog ack failed', err));
-    });
-  }
-
-  if (shutdownBtn) {
-    shutdownBtn.addEventListener('click', () => {
-      if (!_watchdogPending) return;
-      fetch('/api/watchdog/shutdownnow', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        credentials: 'include',
-        body: JSON.stringify({
-          id: _watchdogPending.id,
-          eventId: _watchdogPending.dbEventId
-        })
-      }).catch(err => console.error('watchdog shutdownnow failed', err));
-    });
-  }
-
-  // Subscribe to live broadcasts via the LiveSource (if present).
-  // The liveSource is created lazily in initModeToggle; wire here, and
-  // also on every recreate via a reattach helper.
-  attachWatchdogWebSocket();
-
-  // Initial state load
-  refreshWatchdogStateFromServer();
-}
-
-function attachWatchdogWebSocket() {
-  if (liveSource && typeof liveSource.onWatchdogState === 'function') {
-    liveSource.onWatchdogState((msg) => {
-      // Mark seeded so any still-in-flight GET /api/watchdog/state
-      // response is discarded rather than clobbering this fresher
-      // WS state.
-      _watchdogStateSeeded = true;
-      _watchdogPending = msg.pending || null;
-      if (msg.snapshot) _watchdogSnapshot = msg.snapshot;
-      if (msg.watchdogs) _watchdogMeta = msg.watchdogs;
-      renderWatchdogBanner(_watchdogPending);
-      renderModeEnablement(_watchdogSnapshot.wb || {}, _watchdogCurrentUserRole());
-      renderWatchdogsCard(_watchdogSnapshot, _watchdogMeta, _watchdogCurrentUserRole());
-      if (msg.recent) renderWatchdogHistory(msg.recent);
-      renderCooloffIndicator(_watchdogSnapshot.wb || {});
-    });
-  }
-}
-
-// Re-attach hook: called when the live source is created in modes
-// where the playground switches from sim to live.
-window.__attachWatchdogWS = attachWatchdogWebSocket;
-
-function refreshWatchdogStateFromServer() {
-  fetch('/api/watchdog/state', { credentials: 'include' })
-    .then(r => r.ok ? r.json() : null)
-    .then(state => {
-      if (!state) return;
-      // If a WebSocket broadcast arrived between issuing this GET and
-      // it resolving, the WS state is fresher and we must NOT clobber
-      // it with the now-stale GET response. The watchdog-flow e2e
-      // test "shutdown now button POSTs..." was flaky on CI under
-      // load specifically because of this race: it injected a WS
-      // broadcast immediately after page load, but the GET resolved
-      // shortly after with `pending: null` and reset the banner.
-      if (_watchdogStateSeeded) return;
-      _watchdogStateSeeded = true;
-      _watchdogPending = state.pending || null;
-      _watchdogSnapshot = state.snapshot || { we: {}, wz: {}, wb: {} };
-      _watchdogMeta = state.watchdogs || [];
-      renderWatchdogBanner(_watchdogPending);
-      renderModeEnablement(_watchdogSnapshot.wb || {}, _watchdogCurrentUserRole());
-      renderWatchdogsCard(_watchdogSnapshot, _watchdogMeta, _watchdogCurrentUserRole());
-      renderWatchdogHistory(state.recent || []);
-      renderCooloffIndicator(_watchdogSnapshot.wb || {});
-    })
-    .catch(() => { /* unauth (public browse) or offline — swallow */ });
-}
-
-function renderWatchdogBanner(pending) {
-  _watchdogPending = pending;
-  const banner = document.getElementById('watchdog-banner');
-  if (!banner) return;
-
-  if (!pending) {
-    banner.style.display = 'none';
-    if (_watchdogCountdownTimer) {
-      clearInterval(_watchdogCountdownTimer);
-      _watchdogCountdownTimer = null;
-    }
-    return;
-  }
-
-  banner.style.display = 'block';
-  const title = document.getElementById('watchdog-banner-title');
-  const reason = document.getElementById('watchdog-banner-reason');
-  const replyInput = document.getElementById('watchdog-banner-reply');
-  const meta = _watchdogMeta.find(w => w.id === pending.id);
-  const label = meta ? meta.shortLabel : pending.id;
-  if (title) title.textContent = 'Watchdog fired: ' + label;
-  if (reason) reason.textContent = pending.triggerReason || '';
-  if (replyInput) replyInput.value = '';
-
-  // Disable buttons for readonly users
-  const role = _watchdogCurrentUserRole();
-  const isAdmin = role === 'admin';
-  const snoozeBtn = document.getElementById('watchdog-banner-snooze');
-  const shutdownBtn = document.getElementById('watchdog-banner-shutdown');
-  if (snoozeBtn) snoozeBtn.disabled = !isAdmin;
-  if (shutdownBtn) shutdownBtn.disabled = !isAdmin;
-
-  // Local countdown ticking every second (5 min = 300s from firedAt)
-  if (_watchdogCountdownTimer) clearInterval(_watchdogCountdownTimer);
-  function updateCountdown() {
-    const now = Math.floor(Date.now() / 1000);
-    const remaining = Math.max(0, 300 - (now - pending.firedAt));
-    const m = Math.floor(remaining / 60);
-    const s = remaining % 60;
-    const el = document.getElementById('watchdog-banner-countdown-text');
-    if (el) el.textContent = m + ' min ' + (s < 10 ? '0' : '') + s + ' s';
-  }
-  updateCountdown();
-  _watchdogCountdownTimer = setInterval(updateCountdown, 1000);
-}
-
-function renderCooloffIndicator(wb) {
-  const el = document.getElementById('watchdog-cooloff-indicator');
-  if (!el) return;
-  const now = Math.floor(Date.now() / 1000);
-  const cooloffs = [];
-  WATCHDOG_ALL_MODES.forEach(m => {
-    const entry = wb && wb[m.code];
-    if (entry && entry > now && entry !== WATCHDOG_PERMANENT_SENTINEL) {
-      const remaining = entry - now;
-      const h = Math.floor(remaining / 3600);
-      const mn = Math.floor((remaining % 3600) / 60);
-      cooloffs.push('\u23F8 ' + m.label + ' cooling off — ' + h + 'h ' + mn + 'm remaining');
-    }
-  });
-  if (cooloffs.length === 0) {
-    el.style.display = 'none';
-    el.textContent = '';
-    return;
-  }
-  el.style.display = 'block';
-  el.textContent = cooloffs.join(' · ');
-}
-
-function renderModeEnablement(wb, userRole) {
-  const list = document.getElementById('mode-enablement-list');
-  if (!list) return;
-  const now = Math.floor(Date.now() / 1000);
-  const isAdmin = userRole === 'admin';
-
-  list.innerHTML = '';
-  WATCHDOG_ALL_MODES.forEach(mode => {
-    const row = document.createElement('div');
-    row.className = 'mode-enablement-row';
-    const entry = wb && wb[mode.code];
-
-    let statusLabel;
-    let actionLabel;
-    let actionHandler;
-
-    if (!entry || entry <= now) {
-      statusLabel = '<span class="mode-allowed">\u2022 allowed</span>';
-      actionLabel = 'Disable';
-      actionHandler = () => _watchdogDisableMode(mode.code);
-    } else if (entry === WATCHDOG_PERMANENT_SENTINEL) {
-      statusLabel = '<span class="mode-disabled">\u2715 disabled by user</span>';
-      actionLabel = 'Re-enable';
-      actionHandler = () => _watchdogClearBan(mode.code);
-    } else {
-      const remaining = entry - now;
-      const h = Math.floor(remaining / 3600);
-      const mn = Math.floor((remaining % 3600) / 60);
-      statusLabel = '<span class="mode-cooloff">\u23F8 cool-off — ' + h + 'h ' + mn + 'm</span>';
-      actionLabel = 'Clear cool-off';
-      actionHandler = () => _watchdogClearBan(mode.code);
-    }
-
-    row.innerHTML =
-      '<div class="mode-enablement-label">' + mode.label + '</div>' +
-      '<div class="mode-enablement-status">' + statusLabel + '</div>' +
-      (isAdmin ? '<button class="mode-enablement-action" type="button">' + actionLabel + '</button>' : '');
-
-    if (isAdmin) {
-      const btn = row.querySelector('.mode-enablement-action');
-      if (btn) btn.addEventListener('click', actionHandler);
-    }
-    list.appendChild(row);
-  });
-}
-
-function _watchdogDisableMode(modeCode) {
-  const body = { wb: {} };
-  body.wb[modeCode] = WATCHDOG_PERMANENT_SENTINEL;
-  fetch('/api/device-config', {
-    method: 'PUT',
-    headers: { 'Content-Type': 'application/json' },
-    credentials: 'include',
-    body: JSON.stringify(body)
-  }).catch(err => console.error('disable mode failed', err));
-}
-
-function _watchdogClearBan(modeCode) {
-  const body = { wb: {} };
-  body.wb[modeCode] = 0;
-  fetch('/api/device-config', {
-    method: 'PUT',
-    headers: { 'Content-Type': 'application/json' },
-    credentials: 'include',
-    body: JSON.stringify(body)
-  }).catch(err => console.error('clear ban failed', err));
-}
-
-function renderWatchdogsCard(snapshot, watchdogs, userRole) {
-  const list = document.getElementById('watchdogs-list');
-  if (!list) return;
-  const isAdmin = userRole === 'admin';
-  const now = Math.floor(Date.now() / 1000);
-
-  list.innerHTML = '';
-  if (!watchdogs || watchdogs.length === 0) {
-    list.innerHTML = '<p class="empty-state">No watchdogs configured.</p>';
-    return;
-  }
-  watchdogs.forEach(w => {
-    const row = document.createElement('div');
-    row.className = 'watchdog-row';
-
-    const enabled = snapshot && snapshot.we && snapshot.we[w.id];
-    const snoozeUntil = snapshot && snapshot.wz && snapshot.wz[w.id];
-    const isSnoozed = snoozeUntil && snoozeUntil > now;
-
-    const label = document.createElement('label');
-    label.className = 'watchdog-row-label';
-
-    const checkbox = document.createElement('input');
-    checkbox.type = 'checkbox';
-    checkbox.checked = !!enabled;
-    checkbox.disabled = !isAdmin;
-    checkbox.addEventListener('change', () => {
-      fetch('/api/watchdog/enabled', {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        credentials: 'include',
-        body: JSON.stringify({ id: w.id, enabled: checkbox.checked })
-      }).catch(err => console.error('toggle watchdog failed', err));
-    });
-    label.appendChild(checkbox);
-
-    const text = document.createElement('span');
-    text.innerHTML = '<strong>' + (w.label || w.id) + '</strong> &mdash; ' + (w.mode || '');
-    label.appendChild(text);
-    row.appendChild(label);
-
-    if (isSnoozed) {
-      const snoozeInfo = document.createElement('div');
-      snoozeInfo.className = 'watchdog-snooze-info';
-      const remaining = snoozeUntil - now;
-      const h = Math.floor(remaining / 3600);
-      const mn = Math.floor((remaining % 3600) / 60);
-      snoozeInfo.textContent = '\u23F8 snoozed for ' + h + 'h ' + mn + 'm';
-      row.appendChild(snoozeInfo);
-      if (isAdmin) {
-        const clearBtn = document.createElement('button');
-        clearBtn.className = 'watchdog-clear-snooze';
-        clearBtn.type = 'button';
-        clearBtn.textContent = 'Clear snooze';
-        clearBtn.addEventListener('click', () => {
-          const body = { wz: {} };
-          body.wz[w.id] = 0;
-          fetch('/api/device-config', {
-            method: 'PUT',
-            headers: { 'Content-Type': 'application/json' },
-            credentials: 'include',
-            body: JSON.stringify(body)
-          }).catch(err => console.error('clear snooze failed', err));
-        });
-        row.appendChild(clearBtn);
-      }
-    }
-    list.appendChild(row);
-  });
-}
-
-function renderWatchdogHistory(recent) {
-  const list = document.getElementById('watchdogs-history-list');
-  if (!list) return;
-  list.innerHTML = '';
-  if (!recent || recent.length === 0) {
-    list.innerHTML = '<p class="empty-state">No events yet.</p>';
-    return;
-  }
-  recent.forEach(ev => {
-    const row = document.createElement('div');
-    row.className = 'watchdog-history-row';
-    const when = ev.fired_at ? new Date(ev.fired_at).toLocaleString() : '';
-    const resolution = ev.resolution || 'pending';
-    const safeReason = (ev.trigger_reason || '').replace(/</g, '&lt;');
-    const safeSnooze = (ev.snooze_reason || '').replace(/</g, '&lt;');
-    row.innerHTML =
-      '<div class="history-row-top">' + when + ' &mdash; <code>' + ev.watchdog_id + '</code> &mdash; ' + resolution + '</div>' +
-      '<div class="history-row-reason">' + safeReason + '</div>' +
-      (ev.snooze_reason ? '<div class="history-row-snooze">Snoozed: "' + safeSnooze + '"</div>' : '');
-    list.appendChild(row);
-  });
-}
 
 init();
